requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
Werkzeug==3.0.3
python-dotenv==1.0.1
SQLAlchemy==2.0.36
alembic==1.13.2
pandas==2.2.3
openpyxl==3.1.5
reportlab==4.2.2
requests==2.32.3

